List plugins and delete plugin functionality for devs

+++ server/src/main/java/org/tyit/pnc/repository/PluginRepository.java
@@ -19,7 +19,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 import org.tyit.pnc.model.Plugin;
 
-public interface PluginRepository extends CrudRepository<Plugin, Integer> {
+public interface PluginRepository extends CrudRepository<Plugin, Long> {
 
   @Query("SELECT P FROM Plugin P WHERE LOWER(P.name) LIKE LOWER(CONCAT(?1, '%'))")
   public Iterable<Plugin> findAllByName(String name);
